RAG-Benchmark-master/requirements.txt
langchain
langchain-community
langchain-openai
langchain-weaviate
openai
tiktoken

# Data processing
pandas
numpy
PyPDF2
PyMuPDF

# Utilities
python-dotenv
pydantic
tqdm

# Testing
pytest
# pytest-asyncio
ragas>=0.1.0
datasets>=2.15.0
evaluate>=0.4.0
